requirements_web.txt
# Streamlit Web 应用依赖

# Web 框架
streamlit>=1.28.0

# LlamaIndex 核心依赖
llama-index-core>=0.10.0
llama-index-llms-deepseek>=0.1.0

# 嵌入模型
sentence-transformers>=2.2.0

# 数据处理
python-dotenv>=1.0.0

# 其他依赖
pydantic>=2.0.0
typing-extensions>=4.0.0
